+++ src/components/blog/postList/postList.scss
@@ -1,0 +1,3 @@
+h2.title.postListTitle {
+    margin-bottom: 0.5rem;
+}
