+++ src/templates/category.js
@@ -2,18 +2,16 @@ import React from 'react'
 import Helmet from 'react-helmet'
 import { graphql } from 'gatsby'
 import Layout from '../components/layout/layout'
-import PostList from '../components/blog/postList'
+import PostList from '../components/blog/postList/postList'
 import Pagination from '../components/blog/pagination'
 import Sidebar from '../components/blog/sidebar'
 
 const Category = props => {
   const { data, pageContext } = props
-  const { edges: posts, totalCount } = data.allWordpressPost
+  const { edges: posts } = data.allWordpressPost
   const { title: siteTitle } = data.site.siteMetadata
   const { name: category } = pageContext
-  const title = `${totalCount} post${
-    totalCount === 1 ? '' : 's'
-  } in the “${category}” category`
+  const title = `Posts in category "${category}"`
 
   return (
     <Layout>
